# api/swagger/swagger.yaml
definitions:
...
  models.PasswordUpdate:
    properties:
      new_password:
        example: new_password
        type: string
      new_password_repeat:
        example: new_password
        type: string
      old_password:
        example: password
        type: string
    type: object
  models.Response:
    properties:
      result:
        type: string
    type: object
  models.Tokens:
    properties:
      access:
        type: string
      expires_at:
        type: integer
      refresh:
        type: string
    type: object
  models.UserAuth:
    properties:
      email:
        example: example@example.com
        type: string
      password:
        example: password
        type: string
    type: object
...
info:
  contact: {}
  title: Crypto Volume Finder API
  version: "1.0"
paths:
  /api/user:
    delete:
      description: Delete the authenticated user's account
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete a user account
      tags:
      - users
  /api/user/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and issue tokens if successful
      parameters:
      - description: User login data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserAuth'
      produces:
      - application/json
      responses:
        "200":
          description: New tokens data
          schema:
            $ref: '#/definitions/models.Tokens'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Log in a user
      tags:
      - users
  /api/user/auth/signup:
    post:
      consumes:
      - application/json
      description: |-
        Create a new user account with email and password.
        Returns the access token, the refresh token, and the time when the access token ceases to be valid. After the access token has ceased to be valid, you need to send a request along the path "/api/user/auth/token" to get a new pair of tokens.
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserAuth'
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with tokens data
          schema:
            $ref: '#/definitions/models.Tokens'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Sign up a new user
      tags:
      - users
  /api/user/auth/tokens:
    get:
      description: Retrieve new access and refresh tokens for the authenticated user
      parameters:
      - description: Refresh token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: Successful response with new tokens
          schema:
            $ref: '#/definitions/models.Tokens'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get new tokens
      tags:
      - users
...
  /api/user/update-password:
    put:
      consumes:
      - application/json
      description: |-
        Update the password for the authenticated user.
        Returns the access token, the refresh token, and the time when the access token ceases to be valid. After the access token has ceased to be valid, you need to send a request along the path "/api/user/auth/token" to get a new pair of tokens.
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Passwords data
        in: body
        name: passwords
        required: true
        schema:
          $ref: '#/definitions/models.PasswordUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: New tokens data
          schema:
            $ref: '#/definitions/models.Tokens'
        "400":
          description: Invalid password
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update user password
      tags:
      - users
swagger: "2.0"